Avoid signed/unsigned integer comparison warning

+++ src/world.c
@@ -25,7 +25,7 @@ void world_destroy(world_t *world)
 void world_populate(world_t *world, size_t pop)
 {
 	scatter_t scat; 
-	if (pop == -1) pop = genome_pool_get_count(&world->genomes);
+	if (pop == (size_t)-1) pop = genome_pool_get_count(&world->genomes);
 	scatter_init(&scat, world->width * world->height, pop);
 	while (scatter_has_next(&scat)) {
 		size_t i = scatter_next(&scat, &world->rand);
